Upgrade df version (#1454)
* Upgrade df version

Upgrade df version to 50.3.0 and arrow/parquet to 57.0.0

* update datafusion version in cargo toml

* fix for aarch64 kafka build

* update for kafka aarch64

* update Cargo.toml, remove creating cargo.toml in build

* update arrow version

- Add Docker Buildx and QEMU setup for ARM64 emulation
- Switch to nightly Rust toolchain (required for Rust 1.88.0)
- Configure Cross.toml with ARM64 package installation and library paths
- Set CROSS_NO_WARNINGS=0 to ignore post-install script warnings
- Add CMake environment variables for Kafka builds to find ARM64 libs

---------

+++ src/storage/localfs.rs
@@ -143,7 +143,7 @@ impl ObjectStorage for LocalFS {
                         .modified()
                         .map_err(ObjectStorageError::IoError)?
                         .into(),
-                    size: metadata.len() as usize,
+                    size: metadata.len(),
                     e_tag: None,
                     version: None,
                 };
